Did some cool stuff....idk

+++ pages/UserPage.php
@@ -34,6 +34,7 @@ session_start();
     sessionStorage.setItem("showProfile", "true");
     document.getElementById("item6").style.display = "block";
     document.getElementById("item5").style.display = "none";
+    sessionStorage.setItem("answer", "yes");
 </script>
 
 <div class="container">
@@ -104,23 +105,40 @@ session_start();
                 <h4>Περιγραφή</h4>
                 <label id="text1" for="exampleFormControlTextarea1" class="form-label"></label>
                 <div id="description">
-                <label><?php echo $_SESSION['about'] ?></label>
+                    <label><?php echo $_SESSION['about'] ?></label>
                 </div>
                 <br><br>
             </div>
             <div class="col-md-12">
                 <form method="post" action="../phpScripts/deleteUser.php" id="del_form">
-                <button onclick="disconnect()" type=button class="btn btn-danger">Αποσύνδεση</button>
-                <button onclick="disconnect()" type=submit class="btn btn-danger" form="del_form">Διαγραφή Λογαριασμού</button>
+                    <button onclick="disconnect(1)" type=button class="btn btn-danger">Αποσύνδεση</button>
+                    <button onclick="disconnect(2)" id="del" type=button class="btn btn-danger">Διαγραφή Λογαριασμού
+                    </button>
                 </form>
             </div>
 
             <script>
-                function disconnect() {
-                    window.location = 'mainpage.php'
-                    document.getElementById("item6").style.display = "none";
-                    document.getElementById("item5").style.display = "block";
-                    sessionStorage.setItem("showProfile","false");
+                function disconnect(text) {
+                    if (text === 2) {
+                        if (sessionStorage.getItem("answer") === null || sessionStorage.getItem("answer") === "yes") {
+                            if (confirm('Are you sure you want to delete your account ?')) {
+                                sessionStorage.setItem("answer", "no");
+                                document.getElementById("del").setAttribute('type', 'submit');
+                                document.getElementById("del").click();
+                                window.location = 'mainpage.php'
+                                document.getElementById("item6").style.display = "none";
+                                document.getElementById("item5").style.display = "block";
+                                sessionStorage.setItem("showProfile", "false");
+                            }
+                        }
+                    } else if (text === 1) {
+                        if (confirm('Are you sure you want to disconnect ?')) {
+                            window.location = 'mainpage.php'
+                            document.getElementById("item6").style.display = "none";
+                            document.getElementById("item5").style.display = "block";
+                            sessionStorage.setItem("showProfile", "false");
+                        }
+                    }
                 }
             </script>
